requirements.txt
Django==4.2.4
Pillow==10.0.0
PyUserInput==0.1.9
asgiref==3.7.2
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
funcparserlib==1.0.1
hy==0.27.0
macro==0.0.2
macros==0.0.1
pip==23.2.1
pyperclip==1.5.26
python-gettext==5.0
pytz==2023.3
setuptools==65.5.1
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
wheel==0.38.4